Replace ForceOverwrite with a cmd line option
At the moment the collision behavior each command gets can be set with the
FileCommandFilter which gets configured in the app.config.
We would like to be able to change the collision behavior just by passing
a command line option.

In this PR I added a command line option `-collisionBehavior` which can be
set to `force` or `skip`. Each command then gets this collisionBehavior
mode.