Changed DiffGenerator quite a bit. For now they are both there, with an option to use -n to use the NewDIffGenerator
Some thoughts behind the changes

Rather than run through all items in the other iterator to check for complex equalities, I created maps with keys given their individual uniqueness. This should give much better runtimes for large sets of data.
The ContentItem equality comparer was based on Sitecore serialization. Unicorn paths are not unique in the same way - and databasename is not a part of unicorn paths (and depending on how the source and target was made, not necesarily part of the Sitecore either). I create the map key by Database and ItemId which will always be unique.
I honestly don't see the need for anything but full path for files. Given the windows file/foldername conventions
Simplified the logic quite a bit (did I oversimplify?). There's no longer any cleanup of unnecesarily created commands.
I added some exceptions which should only happen on corrupted/incorrect data. It could easily move on silently and just choose 1. But I personally believe exceptions with some info is the better way to go.
The test example I supplied in the issue was flawed. It was made by hand and the 2 fields in the source had the same ID. The new code dectected and gave info on this. The old did not.